+++ pyi.py
@@ -42,6 +42,66 @@ class TypeVarInfo(NamedTuple):
     name: str
 
 
+# for both typing and typing_extensions
+_BAD_COLLECTIONS_ALIASES = {
+    "Counter": "Counter",
+    "Deque": "deque",
+    "DefaultDict": "defaultdict",
+    "ChainMap": "ChainMap",
+    "OrderedDict": "OrderedDict",
+}
+_BAD_COLLECTIONS_ALIASES = {
+    alias: f'"collections.{cls}"' for alias, cls in _BAD_COLLECTIONS_ALIASES.items()
+}
+
+# Just for typing
+_BAD_BUILTINS_ALIASES = {
+    alias: f'"builtins.{alias.lower()}"'
+    for alias in ("Dict", "Frozenset", "List", "Set", "Tuple", "Type")
+}
+
+# Just for typing_extensions
+_BAD_CONTEXTLIB_ALIASES = {
+    alias: f'"contextlib.Abstract{alias}" or "typing.{alias}"'
+    for alias in ("ContextManager", "AsyncContextManager")
+}
+_BAD_COLLECTIONS_ABC_ALIASES = {
+    alias: f'"collections.abc.{alias}" or "typing.{alias}"'
+    for alias in (
+        "Awaitable",
+        "Coroutine",
+        "AsyncIterable",
+        "AsyncIterator",
+        "AsyncGenerator",
+    )
+}
+_TYPING_NOT_TYPING_EXTENSIONS = {
+    alias: f'"typing.{alias}"'
+    for alias in (
+        "Protocol",
+        "runtime_checkable",
+        "ClassVar",
+        "NewType",
+        "overload",
+        "Text",
+        "NoReturn",
+    )
+}
+
+
+# collections.abc.Set is dealt with separately as special cases
+FORBIDDEN_IMPORTS_MAPPING = {
+    "typing": {**_BAD_BUILTINS_ALIASES, **_BAD_COLLECTIONS_ALIASES},
+    "typing_extensions": {
+        **_BAD_COLLECTIONS_ABC_ALIASES,
+        **_BAD_CONTEXTLIB_ALIASES,
+        **_BAD_COLLECTIONS_ALIASES,
+        **_TYPING_NOT_TYPING_EXTENSIONS,
+    },
+    "collections": {"namedtuple": '"typing.NamedTuple"'},
+}
+
+
 class PyiAwareFlakesChecker(FlakesChecker):
     def deferHandleNode(self, node, parent):
         self.deferFunction(lambda: self.handleNode(node, parent))
@@ -186,6 +246,50 @@ class PyiVisitor(ast.NodeVisitor):
     def in_class(self) -> bool:
         """Determine whether we are inside a `class` statement"""
         return bool(self._class_nesting)
+
+    def _Y022_check(
+        self,
+        node: ast.Attribute | ast.alias,
+        object_name: str,
+        module_name: str,
+        blacklist: dict[str, str],
+    ) -> None:
+        if object_name in blacklist:
+            error_message = Y022.format(
+                good_cls_name=blacklist[object_name],
+                bad_cls_alias=f'"{module_name}.{object_name}"',
+            )
+            self.error(node, error_message)
+
+    def visit_Attribute(self, node: ast.Attribute) -> None:
+        self.generic_visit(node)
+        thing = node.value
+        if not isinstance(thing, ast.Name):
+            return
+        thingname = thing.id
+        if thingname in FORBIDDEN_IMPORTS_MAPPING:
+            self._Y022_check(
+                node=node,
+                object_name=node.attr,
+                module_name=thingname,
+                blacklist=FORBIDDEN_IMPORTS_MAPPING[thingname],
+            )
+
+    def visit_ImportFrom(self, node: ast.ImportFrom) -> None:
+        module_name, imported_objects = node.module, node.names
+        if module_name in FORBIDDEN_IMPORTS_MAPPING:
+            forbidden_imports = FORBIDDEN_IMPORTS_MAPPING[module_name]
+            for obj in imported_objects:
+                self._Y022_check(
+                    node=obj,
+                    object_name=obj.name,
+                    module_name=module_name,
+                    blacklist=forbidden_imports,
+                )
+        elif module_name == "collections.abc":
+            for obj in imported_objects:
+                if obj.name == "Set" and obj.asname != "AbstractSet":
+                    self.error(obj, Y023)
 
     def visit_Assign(self, node: ast.Assign) -> None:
         if self.in_function:
@@ -756,6 +860,11 @@ Y018 = 'Y018 {typevarlike_cls} "{typevar_name}" is not used'
 Y019 = 'Y019 Use "_typeshed.Self" instead of "{typevar_name}"'
 Y020 = "Y020 Quoted annotations should never be used in stubs"
 Y021 = "Y021 Docstrings should not be included in stubs"
+Y022 = "Y022 Use {good_cls_name} instead of {bad_cls_alias}"
+Y023 = (
+    'Y023 Use "from collections.abc import Set as AbstractSet"'
+    'to avoid confusion with "builtins.set"'
+)
 Y092 = "Y092 Top-level attribute must not have a default value"
 Y093 = "Y093 Use typing_extensions.TypeAlias for type aliases"